Migrate SemanticIndentation and BlankLines to TokenStream architecture
This commit completes the migration of all flat transformations to the
TokenStream architecture, following the progressive integration strategy
outlined in Issue #161.

Changes:
- Created SemanticIndentationMapper implementing StreamMapper trait
- Created BlankLinesMapper implementing StreamMapper trait
- Integrated both mappers into indentation and linebased lexer pipelines
- Removed old sem_indentation.rs and blanklines.rs implementations
- Ported all existing tests (13 for SemanticIndentation, 19 for BlankLines)
- Updated detokenizer tests to use full lex() pipeline
- Updated snapshot tests to reflect correct blank line handling

The mappers use the TokenStream architecture to:
- Convert raw Indentation tokens to semantic Indent/Dedent pairs
- Transform consecutive Newline tokens into BlankLine tokens

All 371 tests pass. The reference parser is now 100% TokenStream-integrated,
as all flat transformations have been migrated.

Closes #161
