+++ src/lex/pipeline/mappers/mod.rs
@@ -3,6 +3,10 @@
 //! This module contains concrete implementations of the StreamMapper trait
 //! that perform specific transformations on TokenStreams.
 
+pub mod blank_lines;
 pub mod normalize_whitespace;
+pub mod semantic_indentation;
 
+pub use blank_lines::BlankLinesMapper;
 pub use normalize_whitespace::NormalizeWhitespaceMapper;
+pub use semantic_indentation::SemanticIndentationMapper;
